pyproject.toml
[project]
name = "apollo-agents"
version = "0.1.0"
description = "AI-powered DJ set builder — multi-agent pipeline from tracks to YouTube video"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "ffmpeg-python>=0.2.0",
    "librosa>=0.11.0",
    "moviepy>=2.2.1",
    "numpy>=2.4.2",
    "openai>=1.0.0",
    "Pillow>=10.0.0",
    "pydub>=0.25.1",
    "pyrubberband>=0.4.0",
    "python-dotenv>=1.0.0",
    "soundfile>=0.13.1",
    "sounddevice>=0.5.0",
    "pyrekordbox>=0.4.0",
]

[dependency-groups]
dev = [
    "markdown>=3.10.2",
    "pytest>=9.0.3",
    "weasyprint>=68.1",
]
web = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "email-validator>=2.0.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
